refactor : 레포지토리 네이밍 Cache -> Redis로 변경
